+++ packages/bakit/tsup.config.ts
@@ -1,5 +1,13 @@
 import { makeConfig } from "../../tsup.config.js";
 
-export default makeConfig({
-	entry: ["src/index.ts"],
-});
+export default makeConfig(
+	{
+		entry: ["src/index.ts"],
+	},
+	{
+		entry: {
+			cli: "src/lib/cli/index.ts",
+		},
+		dts: false,
+	},
+);